pyproject.toml
[project]
name = "hops"
version = "0.1.0"
description = "Heterogeneous Optimized Pipeline Simulator"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Jimmy Dai", email = "jimmydai@umich.edu" },
    { name = "Joshua Hsueh", email = "jhsueh@umich.edu" },
    { name = "Olaf Dsouza", email = "olafpd@umich.edu" },
    { name = "Rishith Seelam", email = "rseelam@umich.edu" },
]
dependencies = [
    "numpy",
    "pyyaml",
]

[dependency-groups]
dev = [
    "pytest",
    "matplotlib",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hops"]
